# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iGO - API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/client/addresses": {
            "get": {
                "tags": [
                    "Client: Address"
                ],
                "summary": "Get client addresses",
                "description": "Get list of all addresses from logged in client",
                "operationId": "getClientAddresses",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Client: Address"
                ],
                "summary": "Save client address",
                "description": "Create new client address - Update client address if address_id not null ",
                "operationId": "updateAddressData",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "address_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "address_type_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "address_name": {
                                        "type": "string",
                                        "example": "Trabalho"
                                    },
                                    "line_1": {
                                        "type": "string",
                                        "example": "Address line 1. Eg: Av. Mauris nec dolor, nº 50"
                                    },
                                    "line_2": {
                                        "type": "string",
                                        "example": "Address line 2. Eg: Praceta São João"
                                    },
                                    "county_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "locality": {
                                        "type": "string",
                                        "example": "Bairro"
                                    },
                                    "post_code": {
                                        "type": "string",
                                        "example": "46703"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Angola"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/v1/client/addresses/{id}": {
            "delete": {
                "tags": [
                    "Client: Address"
                ],
                "summary": "Remove address",
                "description": "Remove client address specified in addres id",
                "operationId": "removeClienteAddress",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}